+++ R/pickup.R
@@ -4,7 +4,9 @@
 #' extracted from a given database.
 #'
 #' @name pickup
-#' @param query_file File name of `sql` query
+#' @param query_file File name of `sql` query. Literal query is acceptable, and
+#'   to be recognized as literal query, the input must be a string containing at
+#'   least one new line.
 #' @param config_where Configuration of "where-clause" of the `sql` query. Can
 #'   be a `list` (mostly from the `config.yml` file) or `data.frame`.
 #' @param dsn The data source name of an **ODBC** database connector. See
@@ -29,8 +31,12 @@ pickup <- function(query_file,
     )
   )
   on.exit(DBI::dbDisconnect(con))
-  query <- readLines(query_file, encoding = encoding) |>
-    stringr::str_c(collapse = "\n") |>
+  query <- ifelse(
+    stringr::str_detect(query_file, "\\n"),
+    query_file,
+    readLines(query_file, encoding = encoding) |>
+      stringr::str_c(collapse = "\n")
+  ) |>
     stringr::str_glue(
       .envir = env(
         where_clause = compose_where(config_where)