+++ tests/testthat/test-search_games.R
@@ -4,11 +4,13 @@ test_that("Correctly combine `game_info`", {
     search_games, "pickup",
     function(query_file, ...) tibble::tibble(read.csv(query_file))
   )
-  expect_snapshot(
-    search_games(NULL, query_file = "dummy/test.sql")
+  expect_snapshot_value(
+    search_games(query_file = "dummy/test.sql"),
+    style = "json2"
   )
-  expect_snapshot(
-    search_games(NULL, known_only = FALSE, query_file = "dummy/test.sql")
+  expect_snapshot_value(
+    search_games(known_only = FALSE, query_file = "dummy/test.sql"),
+    style = "json2"
   )
 })